requirements.txt
psutil==5.8.0
speedtest-cli==2.1.3
urllib3==1.26.5
wadllib==1.3.5
progress==1.6
py-cpuinfo==8.0.0 
WMI==1.5.1
pbr==5.8.0
testresources==2.0.1